requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
iso8601==2.1.0
pydantic==2.11.3
pydantic_core==2.33.1
pypika-tortoise==0.5.0
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2